archive/conversational-ai/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
websockets>=12.0

# ASR
openai-whisper>=20231117
faster-whisper>=0.10.0
torch>=2.1.0
torchaudio>=2.1.0
soundfile>=0.12.1
librosa>=0.10.1

# NLU
transformers>=4.35.0
spacy>=3.7.0
scikit-learn>=1.3.0

# TTS
TTS>=0.21.0
pyttsx3>=2.90

# Dialog
redis>=5.0.0

# Utils
numpy>=1.24.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
structlog>=23.2.0

# Monitoring
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0
